\name{ApiPkgLab5-package}
\alias{ApiPkgLab5-package}
\alias{ApiPkgLab5}
\docType{package}
\title{
\packageTitle{ApiPkgLab5}
}
\description{
\packageDescription{ApiPkgLab5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ApiPkgLab5}
\packageIndices{ApiPkgLab5}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ApiPkgLab5}

Maintainer: \packageMaintainer{ApiPkgLab5}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
