% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{crime_api}
\alias{crime_api}
\title{CrimeData}
\usage{
crime_api()
}
\value{
Dataframe
}
\description{
fecting the json data from the given url
}
\details{
FetchingData
}
\references{
"https://data.montgomerycountymd.gov/resource"
}
