% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{extend_base_url}
\alias{extend_base_url}
\title{UrlFilter}
\usage{
extend_base_url(new_limit)
}
\arguments{
\item{new_limit}{The number of rows api will return}
}
\value{
return the character
}
\description{
Adds the filters in the url
}
\details{
AddingLimit
}
