% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_api_data}
\alias{fetch_api_data}
\title{FecthApiData}
\usage{
fetch_api_data(url = get_url(), limit = 200, offset = 0)
}
\arguments{
\item{url}{The url of our api from where we fetch data}

\item{limit}{It tells the number of observation to get from api}

\item{offset}{The number of json pages to load from the api}
}
\value{
It return a dataframe having the data from api and cleaned
}
\description{
fecting the json data from the given url
}
\details{
FetchingData
}
\references{
"https://data.montgomerycountymd.gov/resource"
}
