% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_data.R
\name{get_city_crimes_loc}
\alias{get_city_crimes_loc}
\title{CityCrimeLocations}
\usage{
get_city_crimes_loc(city_name, cities_df)
}
\arguments{
\item{city_name}{The city for which crime locations to filter}

\item{cities_df}{The crimes dataframe with 10 columns}
}
\value{
Returns the matrix with latitude and longitude
}
\description{
The function returns the locations of crime in specific city
}
\details{
CityCordinates
}
