% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_data.R
\name{get_crime_dataframe}
\alias{get_crime_dataframe}
\title{CrimesDataframe}
\usage{
get_crime_dataframe(limit)
}
\arguments{
\item{limit}{The number of observations to be fetched from the api}
}
\value{
Returns the dataframe with observations
}
\description{
This ftech the api data and return dataframe
}
\details{
CrimeData
}
