% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{get_key_names}
\alias{get_key_names}
\title{KeysName}
\usage{
get_key_names()
}
\value{
It returen a character vaector
}
\description{
This function return the name of the keys to extrat from Json file
}
\details{
KeysName
}
